Repair full public smoke regressions
