+++ examples/loopx-project-skill-goal-text-precedence-smoke.py
@@ -17,11 +17,12 @@ def require(condition: bool, message: str) -> None:
 
 def main() -> int:
     text = SKILL.read_text(encoding="utf-8")
+    compact = " ".join(text.split())
     require("Recognized project-local goal-start command:" in text, "goal-start heading missing")
     require("- `/loopx <goal text>`" in text, "goal-text command missing")
     require("- `/loopx`\n" not in text, "bare /loopx should not be listed as project-local fallback")
     require("If there is any non-whitespace text after `/loopx`, it is goal text." in text, "goal text precedence missing")
-    require("do not downgrade the\nrequest into a status or inspection turn" in text, "downgrade guard missing")
+    require("do not downgrade the request into a status or inspection turn" in compact, "downgrade guard missing")
     require("Bare `/loopx` is read/status-first" not in text, "bare /loopx status-first branch should not steer this skill")
     print("loopx-project-skill-goal-text-precedence-smoke ok")
     return 0
