+++ examples/multi-agent-product-recipe-smoke.py
@@ -101,16 +101,11 @@ def main() -> int:
             "the operator runs one command",
             "the generic multi-agent kernel supplies runner, wake, pane-local tick, status",
             "fixed-prompt wake causes each pane to run\nits local A2A tick",
-            "Concrete KNN proof command",
-            "KNN demo: prove partial selection improves over full sort",
-            "`worker_loop.executed_turn_count` is at least `6`",
-            "`run_holdout_eval`, and `write_evaluation_summary`",
-            "role-declared successor todos",
-            "the role skill declares the next todo\ntarget and the kernel validates the target agent before writing state",
-            "baseline metric is `1.0`",
-            "dev metric is `4.0`",
-            "holdout metric is `4.5`",
-            "It is not, by itself, a claim that visible Codex panes authored\nthe research result",
+            "Concrete KNN visible command",
+            "`--preset knn-demo` is the public signal that LoopX should materialize a small\ngenerated KNN benchmark workspace",
+            "The visible panes get `solution.py` as the\nonly editable file",
+            "dev/holdout uplift is absent until visible roles write real public-safe\n  evidence",
+            "it is not a claim that visible Codex panes authored the research result",
         ],
         source=AUTO_RESEARCH_GUIDE,
     )
