+++ examples/skillsbench-task-source-preflight-smoke.py
@@ -417,11 +417,11 @@ def test_reverse_tunnel_app_goal_defaults_apt_bootstrap_fail_fast() -> None:
         update = load_benchmark_run_ledger(ledger)
         case = update["benchmarks"]["skillsbench@1.1"]["cases"]["apt-bootstrap"]
         assert case["latest_decision"]["decision"] == (
-            "baseline_runner_or_setup_repair_required"
+            "baseline_setup_preflight_selection_required"
         ), case
         latest_run = _latest_decision_run(case)
         assert latest_run["failure_class"] == (
-            "skillsbench_runner_setup_blocked_before_agent_rounds"
+            "skillsbench_docker_apt_setup_risk_preflight_blocked"
         )
         assert (
             "skillsbench_docker_apt_setup_risk_preflight_blocked"
